+++ .pydevproject
@@ -1,14 +1,17 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <?eclipse-pydev version="1.0"?><pydev_project>
-        
-    <pydev_pathproperty name="org.python.pydev.PROJECT_SOURCE_PATH">
-                
-        <path>/${PROJECT_DIR_NAME}</path>
             
-    </pydev_pathproperty>
-        
-    <pydev_property name="org.python.pydev.PYTHON_PROJECT_VERSION">python interpreter</pydev_property>
-        
-    <pydev_property name="org.python.pydev.PYTHON_PROJECT_INTERPRETER">python3</pydev_property>
     
+    <pydev_pathproperty name="org.python.pydev.PROJECT_SOURCE_PATH">
+        <path>/${PROJECT_DIR_NAME}</path>
+        <path>/${PROJECT_DIR_NAME}/TeamDrive_DataDownload</path>
+    </pydev_pathproperty>
+            
+    
+    <pydev_property name="org.python.pydev.PYTHON_PROJECT_VERSION">python interpreter</pydev_property>
+            
+    
+    <pydev_property name="org.python.pydev.PYTHON_PROJECT_INTERPRETER">python3</pydev_property>
+        
+
 </pydev_project>
